// src/types/models/TransferEvent.ts
// Auto-generated , DO NOT EDIT
import {Entity, FunctionPropertyNames, FieldsExpression, GetOptions } from "@subql/types-core";
import assert from 'assert';



export type TransferEventProps = Omit<TransferEvent, NonNullable<FunctionPropertyNames<TransferEvent>> | '_name'>;

/*
 * Compat types allows for support of alternative `id` types without refactoring the node
 */
type CompatTransferEventProps = Omit<TransferEventProps, 'id'> & { id: string; };
type CompatEntity = Omit<Entity, 'id'> & { id: string; };

export class TransferEvent implements CompatEntity {

    constructor(
        
        id: string,
        blockHeight: bigint,
        tag: string,
        transactionHash: string,
    ) {
        this.id = id;
        this.blockHeight = blockHeight;
        this.tag = tag;
        this.transactionHash = transactionHash;
        
    }

    public id: string;
    public blockHeight: bigint;
    public tag: string;
    public transactionHash: string;
    

    get _name(): string {
        return 'TransferEvent';
    }

    async save(): Promise<void> {
        const id = this.id;
        assert(id !== null, "Cannot save TransferEvent entity without an ID");
        await store.set('TransferEvent', id.toString(), this as unknown as CompatTransferEventProps);
    }

    static async remove(id: string): Promise<void> {
        assert(id !== null, "Cannot remove TransferEvent entity without an ID");
        await store.remove('TransferEvent', id.toString());
    }

    static async get(id: string): Promise<TransferEvent | undefined> {
        assert((id !== null && id !== undefined), "Cannot get TransferEvent entity without an ID");
        const record = await store.get('TransferEvent', id.toString());
        if (record) {
            return this.create(record as unknown as TransferEventProps);
        } else {
            return;
        }
    }


    /**
     * Gets entities matching the specified filters and options.
     *
     * ⚠️ This function will first search cache data followed by DB data. Please consider this when using order and offset options.⚠️
     * */
    static async getByFields(filter: FieldsExpression<TransferEventProps>[], options: GetOptions<TransferEventProps>): Promise<TransferEvent[]> {
        const records = await store.getByFields<CompatTransferEventProps>('TransferEvent', filter  as unknown as FieldsExpression<CompatTransferEventProps>[], options as unknown as GetOptions<CompatTransferEventProps>);
        return records.map(record => this.create(record as unknown as TransferEventProps));
    }

    static create(record: TransferEventProps): TransferEvent {
        assert(record.id !== undefined && record.id !== null, "id must be provided");
        const entity = new this(
            record.id,
            record.blockHeight,
            record.tag,
            record.transactionHash,
        );
        Object.assign(entity,record);
        return entity;
    }
}